Generalize amount options component & in Staking
