+++ src/components/RowAmountOption/RowAmountOption.less
@@ -1,0 +1,20 @@
+.row-amount-option.ant-form-item {
+  display: flex;
+  flex-direction: row !important;
+  justify-content: space-between !important;
+  width: 100%;
+
+  .ant-btn {
+    width: 23% !important;
+    height: 30px;
+    margin: 5px 0;
+    padding: 4px !important;
+    border: none !important;
+    border-radius: 25px;
+    box-shadow: 0 0 2px 0 #c2c2c2;
+
+    span {
+      font-size: 14px !important;
+    }
+  }
+}
