+++ src/pages/bridge/components/CronosBridgeForm.less
@@ -18,30 +18,6 @@
       }
     }
 
-    .row-amount-option {
-      height: 40px;
-
-      .ant-form-item {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-
-        .ant-btn {
-          width: 23%;
-          height: 30px;
-          margin: 5px 0;
-          padding: 4px !important;
-          border: none;
-          border-radius: 25px;
-          box-shadow: 0 0 2px 0 #c2c2c2;
-
-          span {
-            font-size: 14px !important;
-          }
-        }
-      }
-    }
-
     .review-container {
       margin: 10px 0 20px 0;
       padding: 20px;
