+++ src/pages/staking/staking.less
@@ -27,7 +27,7 @@
 
       .available {
         position: relative;
-        margin: 5px 0 20px 0;
+        margin: 5px 0 0 0;
 
         span {
           color: @font-color;
